refactor: generalize parallel executor with Go generics
Extract executeDiffJobsInParallel into a generic executeJobsInParallel[T any]
in parallel.go to enable reuse across commands.
